UNG-1358 Organizing light sdk
