+++ sdk/src/main/scala/com/ubirch/utils/Paths.scala
@@ -1,0 +1,10 @@
+package com.ubirch.utils
+
+object Paths {
+
+  final val TOKEN_PUBLIC_KEY_PATH = "token.tokenPublicKey"
+  final val VALID_ISSUER_PATH = "token.issuer"
+  final val VALID_AUDIENCE_PATH = "token.audience"
+  final val VALID_ROLES_PATH = "token.roles"
+
+}
